Version 1.1.0:- Big Changes Messaging feature bugs fixed, messages are loading, and can be sent. API Called every 3 seconds to imitate real time  connection.
Message window has a scroll to bottom FAB which is as of now working but can be made more workable. Will be working on it .

Will be working on other message features (like a message, reply to a specific message and all)

As of version version 1.1.0 Home, Network, Messaging is working.

That's it for this patch.